reorganization of constraints
