+++ py4DSTEM/process/phase/iterative_simultaneous_ptychography.py
@@ -20,7 +20,6 @@ from py4DSTEM.io import DataCube
 from py4DSTEM.process.phase.iterative_base_class import PhaseReconstruction
 from py4DSTEM.process.phase.utils import (
     ComplexProbe,
-    estimate_global_transformation_ransac,
     fft_shift,
     generate_batches,
     polar_aliases,
@@ -1846,94 +1845,6 @@ class SimultaneousPtychographicReconstruction(PhaseReconstruction):
 
         return current_object, current_probe
 
-    def _position_correction(
-        self,
-        current_object,
-        shifted_probes,
-        overlap,
-        amplitudes,
-        current_positions,
-        positions_step_size,
-    ):
-        """
-        Position correction using estimated intensity gradient.
-
-        Parameters
-        --------
-        current_object: np.ndarray
-            Current object estimate
-        shifted_probes:np.ndarray
-            fractionally-shifted probes
-        overlap: np.ndarray
-            object * probe overlap
-        amplitudes: np.ndarray
-            Measured amplitudes
-        current_positions: np.ndarray
-            Current positions estimate
-        positions_step_size: float
-            Positions step size
-
-        Returns
-        --------
-        updated_positions: np.ndarray
-            Updated positions estimate
-        """
-
-        xp = self._xp
-
-        electrostatic_obj, _ = current_object
-
-        obj_rolled_x_patches = electrostatic_obj[
-            (self._vectorized_patch_indices_row + 1) % self._object_shape[0],
-            self._vectorized_patch_indices_col,
-        ]
-        obj_rolled_y_patches = electrostatic_obj[
-            self._vectorized_patch_indices_row,
-            (self._vectorized_patch_indices_col + 1) % self._object_shape[1],
-        ]
-
-        overlap_fft = xp.fft.fft2(overlap[0])
-
-        exit_waves_dx_fft = overlap_fft - xp.fft.fft2(
-            obj_rolled_x_patches * shifted_probes
-        )
-        exit_waves_dy_fft = overlap_fft - xp.fft.fft2(
-            obj_rolled_y_patches * shifted_probes
-        )
-
-        overlap_fft_conj = xp.conj(overlap_fft)
-        estimated_intensity = xp.abs(overlap_fft) ** 2
-        measured_intensity = amplitudes[0] ** 2
-
-        flat_shape = (overlap[0].shape[0], -1)
-        difference_intensity = (measured_intensity - estimated_intensity).reshape(
-            flat_shape
-        )
-
-        partial_intensity_dx = 2 * xp.real(
-            exit_waves_dx_fft * overlap_fft_conj
-        ).reshape(flat_shape)
-        partial_intensity_dy = 2 * xp.real(
-            exit_waves_dy_fft * overlap_fft_conj
-        ).reshape(flat_shape)
-
-        coefficients_matrix = xp.dstack((partial_intensity_dx, partial_intensity_dy))
-
-        # positions_update = xp.einsum(
-        #    "idk,ik->id", xp.linalg.pinv(coefficients_matrix), difference_intensity
-        # )
-
-        coefficients_matrix_T = coefficients_matrix.conj().swapaxes(-1, -2)
-        positions_update = (
-            xp.linalg.inv(coefficients_matrix_T @ coefficients_matrix)
-            @ coefficients_matrix_T
-            @ difference_intensity[..., None]
-        )
-
-        current_positions -= positions_step_size * positions_update[..., 0]
-
-        return current_positions
-
     def _warmup_object_threshold_constraint(self, current_object, pure_phase_object):
         """
         Ptychographic threshold constraint.
@@ -2006,7 +1917,7 @@ class SimultaneousPtychographicReconstruction(PhaseReconstruction):
 
         return current_object
 
-    def _warmup_object_smoothness_constraint(
+    def _warmup_object_gaussian_constraint(
         self, current_object, gaussian_filter_sigma, pure_phase_object
     ):
         """
@@ -2045,7 +1956,7 @@ class SimultaneousPtychographicReconstruction(PhaseReconstruction):
 
         return current_object
 
-    def _object_smoothness_constraint(
+    def _object_gaussian_constraint(
         self, current_object, gaussian_filter_sigma, pure_phase_object
     ):
         """
@@ -2206,114 +2117,6 @@ class SimultaneousPtychographicReconstruction(PhaseReconstruction):
 
         return shifted_probe
 
-    def _probe_fourier_amplitude_constraint(self, current_probe):
-        """
-        Ptychographic probe Fourier-amplitude constraint.
-        Used for fixing the probe's amplitude in Fourier space.
-
-        Parameters
-        --------
-        current_probe: np.ndarray
-            Current probe estimate
-
-        Returns
-        --------
-        constrained_probe: np.ndarray
-            Fourier-amplitude constrained probe estimate
-        """
-        xp = self._xp
-
-        current_probe_fft = xp.fft.fft2(current_probe)
-        current_probe_fft_phase = xp.angle(current_probe_fft)
-
-        constrained_probe_fft = self._probe_initial_fft_amplitude * xp.exp(
-            1j * current_probe_fft_phase
-        )
-        constrained_probe = xp.fft.ifft2(constrained_probe_fft)
-
-        return constrained_probe
-
-    def _probe_finite_support_constraint(self, current_probe):
-        """
-        Ptychographic probe support constraint.
-        Used for penalizing focused probes to replicate sample periodicity.
-
-        Parameters
-        --------
-        current_probe: np.ndarray
-            Current probe estimate
-
-        Returns
-        --------
-        constrained_probe: np.ndarray
-            Finite-support constrained probe estimate
-        """
-
-        return current_probe * self._probe_support_mask
-
-    def _positions_center_of_mass_constraint(self, current_positions):
-        """
-        Ptychographic position center of mass constraint.
-        Additionally updates vectorized indices used in _overlap_projection.
-
-        Parameters
-        ----------
-        current_positions: np.ndarray
-            Current positions estimate
-
-        Returns
-        --------
-        constrained_positions: np.ndarray
-            CoM constrained positions estimate
-        """
-        xp = self._xp
-
-        current_positions -= xp.mean(current_positions, axis=0) - self._positions_px_com
-        self._positions_px_fractional = current_positions - xp.round(current_positions)
-
-        self._set_vectorized_patch_indices()
-
-        return current_positions
-
-    def _positions_affine_transformation_constraint(
-        self, initial_positions, current_positions
-    ):
-        """
-        Constrains the updated positions to be an affine transformation of the initial scan positions,
-        composing of two scale factors, a shear, and a rotation angle.
-
-        Uses RANSAC to estimate the global transformation robustly.
-        Stores the AffineTransformation in self._tf.
-
-        Parameters
-        ----------
-        initial_positions: np.ndarray
-            Initial scan positions
-        current_positions: np.ndarray
-            Current positions estimate
-
-        Returns
-        -------
-        constrained_positions: np.ndarray
-            Affine-transform constrained positions estimate
-        """
-
-        xp = self._xp
-
-        tf, _ = estimate_global_transformation_ransac(
-            positions0=initial_positions,
-            positions1=current_positions,
-            origin=self._positions_px_com,
-            translation_allowed=True,
-            min_sample=self._num_diffraction_patterns // 10,
-            xp=xp,
-        )
-
-        self._tf = tf
-        current_positions = tf(initial_positions, origin=self._positions_px_com, xp=xp)
-
-        return current_positions
-
     def _constraints(
         self,
         current_object,
@@ -2374,11 +2177,11 @@ class SimultaneousPtychographicReconstruction(PhaseReconstruction):
 
         if gaussian_filter:
             if warmup_iteration:
-                current_object = self._warmup_object_smoothness_constraint(
+                current_object = self._warmup_object_gaussian_constraint(
                     current_object, gaussian_filter_sigma, pure_phase_object
                 )
             else:
-                current_object = self._object_smoothness_constraint(
+                current_object = self._object_gaussian_constraint(
                     current_object, gaussian_filter_sigma, pure_phase_object
                 )
 
@@ -2772,10 +2575,10 @@ class SimultaneousPtychographicReconstruction(PhaseReconstruction):
                 # position correction
                 if a0 >= fix_positions_iter:
                     positions_px[start:end] = self._position_correction(
-                        self._object,
+                        self._object[0],
                         shifted_probes,
-                        overlap,
-                        amplitudes,
+                        overlap[0],
+                        amplitudes[0],
                         self._positions_px,
                         positions_step_size,
                     )
